+++ lib/shell.c
@@ -15,7 +15,7 @@
 #include <errno.h>
 #include "prototypes.h"
 #include "defines.h"
-#include "string/sprintf.h"
+#include "string/sprintf/snprintf.h"
 
 
 extern char **newenvp;
