+++ src/chage.c
@@ -34,7 +34,7 @@
 #include "pwio.h"
 #include "shadowio.h"
 #include "shadowlog.h"
-#include "string/sprintf.h"
+#include "string/sprintf/snprintf.h"
 #include "string/strftime.h"
 #include "string/strtcpy.h"
 #include "time/day_to_str.h"
