+++ src/gpasswd.c
@@ -35,7 +35,7 @@
 /*@-exitarg@*/
 #include "exitcodes.h"
 #include "shadowlog.h"
-#include "string/sprintf.h"
+#include "string/sprintf/snprintf.h"
 #include "string/strtcpy.h"
