+++ src/groupmod.c
@@ -39,8 +39,8 @@
 #include "sgroupio.h"
 #endif
 #include "shadowlog.h"
+#include "string/sprintf/stpeprintf.h"
 #include "string/stpecpy.h"
-#include "string/stpeprintf.h"
 
 
 /*
