+++ src/login.c
@@ -38,7 +38,7 @@
 /*@-exitarg@*/
 #include "exitcodes.h"
 #include "shadowlog.h"
-#include "string/sprintf.h"
+#include "string/sprintf/snprintf.h"
 #include "string/strftime.h"
 #include "string/strtcpy.h"
