+++ src/useradd.c
@@ -65,7 +65,8 @@
 #include "tcbfuncs.h"
 #endif
 #include "shadowlog.h"
-#include "string/sprintf.h"
+#include "string/sprintf/snprintf.h"
+#include "string/sprintf/xasprintf.h"
 
 
 #ifndef SKEL_DIR
