+++ src/vipw.c
@@ -44,7 +44,8 @@
 #include "tcbfuncs.h"
 #endif				/* WITH_TCB */
 #include "shadowlog.h"
-#include "string/sprintf.h"
+#include "string/sprintf/snprintf.h"
+#include "string/sprintf/xasprintf.h"
 
 
 #define MSG_WARN_EDIT_OTHER_FILE _( \
